Fixed loading of Tribler core process
